moved UIWait and UIFind outside of the factories package and made their constructors public
