+++ Framework/jmaqs-selenium/src/main/java/com/magenic/jmaqs/selenium/SeleniumWait.java
@@ -35,7 +35,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
  * Selenium waiter This is the tool-class used for waiting. It can be used to wait for many
  * circumstances (element exist, not exist, be visible, to equal a desired value, etc..
  * @deprecated {@link SeleniumWait} is no longer supported. Use {@link com.magenic.jmaqs.selenium.factories.UIWaitFactory UIWaitFactory} 
- *             to create {@link com.magenic.jmaqs.selenium.factories.UIWait UIWait} objects instead.
+ *             to create {@link UIWait UIWait} objects instead.
  */
 @Deprecated
 public class SeleniumWait {
