+++ Framework/jmaqs-selenium/src/main/java/com/magenic/jmaqs/selenium/factories/UIFindFactory.java
@@ -1,5 +1,6 @@
 package com.magenic.jmaqs.selenium.factories;
 
+import com.magenic.jmaqs.selenium.UIFind;
 import org.openqa.selenium.SearchContext;
 
 /**
