+++ Framework/jmaqs-selenium/src/test/java/com/magenic/jmaqs/selenium/UIWaitFactoryUnitTest.java
@@ -7,7 +7,6 @@ import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import com.magenic.jmaqs.selenium.factories.UIWait;
 import com.magenic.jmaqs.selenium.factories.UIWaitFactory;
 import com.magenic.jmaqs.selenium.unittestpagemodel.PageElementsPageModel;
 import com.magenic.jmaqs.utilities.helper.TestCategories;
